Add and Norm: redirection.c / redirection_utils.c
