+++ oh_my_minishell/Makefile
@@ -9,7 +9,7 @@ UTILS		= ./libft/ft_memset.c ./libft/ft_bzero.c  ./libft/ft_memcpy.c  ./libft/ft
 								./gnl/get_next_line.c ./gnl/get_next_line_utils.c
 SRCS		= main.c execute_commands.c get_commands_from_gnl.c print_free.c utils_jikang.c pipe.c signal.c init.c exit.c vector.c refine_line_a.c refine_line_b.c refine_line_c.c\
 				 check_command.c execute_echo.c execute_pwd.c execute_exit.c execute_env.c execute_unset.c execute_export.c execute_cd.c execute_path.c execute_is_dir_file.c \
-				 	is_execve.c redirection.c ft_strsemi.c msg_error.c check_command_utils.c pipe_utils.c redirection_execute_part.c execute_path_utils.c execute_cd_utils.c
+				 	is_execve.c redirection.c ft_strsemi.c msg_error.c check_command_utils.c pipe_utils.c redirection_execute_part.c execute_path_utils.c execute_cd_utils.c redirection_utils.c
 OBJS		= $(SRCS:.c=.o)
 OBJS_UTILS	= $(UTILS:.c=.o)
 CC			= gcc
